pyproject.toml
[tool.poetry]
name = "guest-be"
version = "0.1.0"
description = ""
authors = ["DongwookKim0823 <dwkim0823@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-environ = "^0.11.2"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
attrs = "^23.2.0"
requests = "^2.32.2"
django-storages = {extras = ["aiobotocore"], version = "^1.14.2"}
djangorestframework = "^3.15.1"
django-cors-headers = "^4.2.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
djangorestframework-simplejwt = "5.3.0"
djangorestframework-camel-case = "^1.4.2"
boto3 = "^1.34.34"
django-filter = "^24.1"
dependency-injector-fork = "^4.42.1"
sentry-sdk = {extras = ["django"], version = "^2.3.1"}
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
pytz = "^2024.1"
django-concurrency = "^2.5"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-django = "^4.7.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.4.0"

factory-boy = "^3.3.0"
faker = "^19.6.1"

django-debug-toolbar = "^4.2.0"

mypy = "^1.5.1"

types-pillow = "^10.1.0.2"
django-stubs = "^5.0.0"
djangorestframework-stubs = "^3.14.2"
celery-stubs = "^0.1.3"

flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^24.3.0"
pre-commit = "^3.4.0"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 120
target-version = ['py311']
force-exclude = 'migrations'

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["*/migrations/*", "*/config/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

